core/requirements.txt
asgiref==3.7.2
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
click==8.1.5
colorama==0.4.6
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
easyparse==1.0.6
Flask==2.3.2
gevent==23.7.0
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mutagen==1.46.0
Pillow==10.0.0
pip==23.2
psycopg2==2.9.6
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
websockets==11.0.3
Werkzeug==2.3.6
yt-dlp==2023.7.6
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
gunicorn==21.2.0